voice assitance pharma/backend/requirements.txt
# PharmaGenie Voice Agent - Backend
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.5
httpx>=0.26
python-dotenv>=1.0
openai>=1.12
